refactor: decrease class count
